refactor: separate profile content from page.tsx
